console: Refactor error view into global utility component
