+++ pkg/webui/console/views/app/index.js
@@ -29,6 +29,7 @@ import { withEnv } from '@ttn-lw/lib/components/env'
 import ErrorView from '@ttn-lw/lib/components/error-view'
 import ScrollToTop from '@ttn-lw/lib/components/scroll-to-top'
 import WithAuth from '@ttn-lw/lib/components/with-auth'
+import FullViewError, { FullViewErrorInner } from '@ttn-lw/lib/components/full-view-error'
 
 import Header from '@console/containers/header'
 
@@ -38,7 +39,6 @@ import Gateways from '@console/views/gateways'
 import Organizations from '@console/views/organizations'
 import Admin from '@console/views/admin'
 import User from '@console/views/user'
-import FullViewError, { FullViewErrorInner } from '@console/views/error'
 
 import PropTypes from '@ttn-lw/lib/prop-types'
 import dev from '@ttn-lw/lib/dev'
@@ -129,10 +129,12 @@ class ConsoleApp extends React.PureComponent {
       },
     } = this.props
 
+    const header = <Header className={style.header} />
+
     if (pageData && pageData.error) {
       return (
         <ConnectedRouter history={history}>
-          <FullViewError error={pageData.error} />
+          <FullViewError error={pageData.error} header={header} />
         </ConnectedRouter>
       )
     }
@@ -149,7 +151,7 @@ class ConsoleApp extends React.PureComponent {
                 defaultTitle={siteName}
               />
               <div id="modal-container" />
-              <Header className={style.header} />
+              {header}
               <main className={style.main}>
                 <WithAuth
                   user={user}
