+++ pkg/webui/console/views/application-integrations-pubsubs/index.js
@@ -26,7 +26,7 @@ import withFeatureRequirement from '@console/lib/components/with-feature-require
 import ApplicationPubsubEdit from '@console/views/application-integrations-pubsub-edit'
 import ApplicationPubsubAdd from '@console/views/application-integrations-pubsub-add'
 import ApplicationPubsubsList from '@console/views/application-integrations-pubsubs-list'
-import SubViewError from '@console/views/error/sub-view'
+import SubViewError from '@console/views/sub-view-error'
 
 import PropTypes from '@ttn-lw/lib/prop-types'
 import sharedMessages from '@ttn-lw/lib/shared-messages'
