+++ pkg/webui/console/views/sub-view-error/index.js
