+++ pkg/webui/lib/components/full-view-error/connect.js
