+++ pkg/webui/lib/components/full-view-error/error.js
