+++ pkg/webui/lib/components/full-view-error/index.js
